Move AuthLogo to plugin
